cf_source/requirements.txt
pdfplumber>=0.11.0
pymongo>=4.7.1
google-cloud-storage>=2.19.0
functions-framework>=3.8.0
